+++ allensdk/brain_observatory/behavior/data_objects/__init__.py
@@ -1,4 +1,4 @@
-from allensdk.brain_observatory.behavior.data_objects._base._data_object_abc import DataObject  # noqa: E501, F401
+from allensdk.brain_observatory.behavior.data_objects.base._data_object_abc import DataObject  # noqa: E501, F401
 from allensdk.brain_observatory.behavior.data_objects.metadata\
     .behavior_metadata.behavior_session_id import BehaviorSessionId  # noqa: E501, F401
 from allensdk.brain_observatory.behavior.data_objects.stimulus_timestamps.stimulus_timestamps import StimulusTimestamps  # noqa: E501, F401
