+++ allensdk/brain_observatory/behavior/data_objects/base/readable_interfaces/stimulus_file_readable_interface.py
