+++ allensdk/brain_observatory/behavior/data_objects/base/writable_interfaces/nwb_writable_interface.py
