+++ allensdk/brain_observatory/behavior/data_objects/metadata/behavior_metadata/behavior_session_id.py
@@ -3,16 +3,16 @@ from pynwb import NWBFile
 from cachetools import cached, LRUCache
 from cachetools.keys import hashkey
 
-from allensdk.brain_observatory.behavior.data_objects._base\
+from allensdk.brain_observatory.behavior.data_objects.base\
     .readable_interfaces.json_readable_interface import \
     JsonReadableInterface
-from allensdk.brain_observatory.behavior.data_objects._base.readable_interfaces\
+from allensdk.brain_observatory.behavior.data_objects.base.readable_interfaces\
     .lims_readable_interface import \
     LimsReadableInterface
-from allensdk.brain_observatory.behavior.data_objects._base\
+from allensdk.brain_observatory.behavior.data_objects.base\
     .readable_interfaces.nwb_readable_interface import \
     NwbReadableInterface
-from allensdk.brain_observatory.behavior.data_objects._base\
+from allensdk.brain_observatory.behavior.data_objects.base\
     .writable_interfaces.json_writable_interface import \
     JsonWritableInterface
 from allensdk.internal.api import PostgresQueryMixin
