+++ allensdk/brain_observatory/behavior/data_objects/metadata/behavior_metadata/stimulus_frame_rate.py
@@ -2,10 +2,10 @@ from pynwb import NWBFile
 
 from allensdk.brain_observatory.behavior.data_objects import DataObject, \
     StimulusTimestamps
-from allensdk.brain_observatory.behavior.data_objects._base\
+from allensdk.brain_observatory.behavior.data_objects.base\
     .readable_interfaces.nwb_readable_interface import \
     NwbReadableInterface
-from allensdk.brain_observatory.behavior.data_objects._base.readable_interfaces\
+from allensdk.brain_observatory.behavior.data_objects.base.readable_interfaces\
     .stimulus_file_readable_interface \
     import \
     StimulusFileReadableInterface
