+++ allensdk/brain_observatory/behavior/data_objects/metadata/subject_metadata/mouse_id.py
@@ -1,16 +1,16 @@
 from pynwb import NWBFile
 
 from allensdk.brain_observatory.behavior.data_objects import DataObject
-from allensdk.brain_observatory.behavior.data_objects._base\
+from allensdk.brain_observatory.behavior.data_objects.base\
     .readable_interfaces.json_readable_interface import \
     JsonReadableInterface
-from allensdk.brain_observatory.behavior.data_objects._base.readable_interfaces\
+from allensdk.brain_observatory.behavior.data_objects.base.readable_interfaces\
     .lims_readable_interface import \
     LimsReadableInterface
-from allensdk.brain_observatory.behavior.data_objects._base\
+from allensdk.brain_observatory.behavior.data_objects.base\
     .readable_interfaces.nwb_readable_interface import \
     NwbReadableInterface
-from allensdk.brain_observatory.behavior.data_objects._base\
+from allensdk.brain_observatory.behavior.data_objects.base\
     .writable_interfaces.json_writable_interface import \
     JsonWritableInterface
 from allensdk.internal.api import PostgresQueryMixin
