+++ allensdk/brain_observatory/behavior/data_objects/running_speed/running_speed.py
@@ -10,19 +10,19 @@ import pandas as pd
 from pynwb import NWBFile, ProcessingModule
 from pynwb.base import TimeSeries
 
-from allensdk.brain_observatory.behavior.data_objects._base\
+from allensdk.brain_observatory.behavior.data_objects.base\
     .readable_interfaces.json_readable_interface import \
     JsonReadableInterface
-from allensdk.brain_observatory.behavior.data_objects._base\
+from allensdk.brain_observatory.behavior.data_objects.base\
     .readable_interfaces.lims_readable_interface import \
     LimsReadableInterface
-from allensdk.brain_observatory.behavior.data_objects._base\
+from allensdk.brain_observatory.behavior.data_objects.base\
     .readable_interfaces.nwb_readable_interface import \
     NwbReadableInterface
-from allensdk.brain_observatory.behavior.data_objects._base\
+from allensdk.brain_observatory.behavior.data_objects.base\
     .writable_interfaces.json_writable_interface import \
     JsonWritableInterface
-from allensdk.brain_observatory.behavior.data_objects._base\
+from allensdk.brain_observatory.behavior.data_objects.base\
     .writable_interfaces.nwb_writable_interface import \
     NwbWritableInterface
 from allensdk.core.exceptions import DataFrameIndexError
